+++ public/actions/AboutOurHoneymoonActions.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import BaseActions from './BaseActions';
-import api from '../api/aboutOurHoneymoon.api';
-import { ABOUT_OUR_HONEYMOON as key } from '../constants/keys.constants';
+import api from '../api/AboutOurHoneymoonApi';
+import { ABOUT_OUR_HONEYMOON as key } from '../constants/KeyConstants';
 
 class AboutOurHoneymoonActions extends BaseActions {
     constructor() {
