+++ public/actions/AboutUsActions.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import BaseActions from './BaseActions';
-import api from '../api/aboutUs.api';
-import { ABOUT_US as key } from '../constants/keys.constants';
+import api from '../api/AboutUsApi';
+import { ABOUT_US as key } from '../constants/KeyConstants';
 
 class AboutUsActions extends BaseActions {
     constructor() {
