+++ public/actions/LocalFlavourActions.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import BaseActions from './BaseActions';
-import api from '../api/localFlavour.api';
-import { LOCAL_FLAVOUR as key } from '../constants/keys.constants';
+import api from '../api/LocalFlavourApi';
+import { LOCAL_FLAVOUR as key } from '../constants/KeyConstants';
 
 class LocalFlavourActions extends BaseActions {
     constructor() {
