+++ public/actions/RsvpActions.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import BaseActions from './BaseActions';
-import api from '../api/rsvp.api';
-import { RSVP as key } from '../constants/keys.constants';
+import api from '../api/RsvpApi';
+import { RSVP as key } from '../constants/KeyConstants';
 
 class RsvpActions extends BaseActions {
     constructor() {
