+++ public/actions/WeddingPartyMemberActions.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import BaseActions from './BaseActions';
 import api from '../api/WeddingPartyMembersApi';
-import { WEDDING_PARTY_MEMBER as key } from '../constants/keys.constants';
+import { WEDDING_PARTY_MEMBER as key } from '../constants/KeyConstants';
 
 class WeddingPartyMemberActions extends BaseActions {
     constructor() {
