+++ public/components/AboutOurDay/AboutOurDayPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import AboutOurDayStore from '../../stores/AboutOurDayStore';
 import AboutOurDayActions from '../../actions/AboutOurDayActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { ABOUT_OUR_DAY as key } from '../../constants/keys.constants';
+import { ABOUT_OUR_DAY as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
