+++ public/components/AboutOurHoneymoon/AboutOurHoneymoonPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import AboutOurHoneymoonStore from '../../stores/AboutOurHoneymoonStore';
 import AboutOurHoneymoonActions from '../../actions/AboutOurHoneymoonActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { ABOUT_OUR_HONEYMOON as key } from '../../constants/keys.constants';
+import { ABOUT_OUR_HONEYMOON as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
