+++ public/components/AboutUs/AboutUsPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import AboutUsStore from '../../stores/AboutUsStore';
 import AboutUsActions from '../../actions/AboutUsActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { ABOUT_US as key } from '../../constants/keys.constants';
+import { ABOUT_US as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
