+++ public/components/LocalFlavour/LocalFlavourPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import LocalFlavourStore from '../../stores/LocalFlavourStore';
 import LocalFlavourActions from '../../actions/LocalFlavourActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { LOCAL_FLAVOUR as key } from '../../constants/keys.constants';
+import { LOCAL_FLAVOUR as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
