+++ public/components/OnTheDay/OnTheDayPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import OnTheDayStore from '../../stores/OnTheDayStore';
 import OnTheDayActions from '../../actions/OnTheDayActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { ON_THE_DAY as key } from '../../constants/keys.constants';
+import { ON_THE_DAY as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
