+++ public/components/Rsvp/RsvpPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import RsvpStore from '../../stores/RsvpStore';
 import RsvpActions from '../../actions/RsvpActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { RSVP as key } from '../../constants/keys.constants';
+import { RSVP as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
