+++ public/components/WeddingPlaylist/WeddingPlaylistPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import WeddingPlaylistStore from '../../stores/WeddingPlaylistStore';
 import WeddingPlaylistActions from '../../actions/WeddingPlaylistActions';
 import MarkdownEditorPage from '../common/MarkdownEditorPage';
-import { WEDDING_PLAYLIST as key } from '../../constants/keys.constants';
+import { WEDDING_PLAYLIST as key } from '../../constants/KeyConstants';
 
 export default function AboutOurDayPage() {
     return (
