+++ public/constants/KeyConstants.js
@@ -6,3 +6,4 @@ export const ON_THE_DAY = 'onTheDay';
 export const RSVP = 'rsvp';
 export const WEDDING_PLAYLIST = 'weddingPlaylist';
 export const WEDDING_PARTY_MEMBER = 'member';
+export const COVER = 'cover';
