+++ public/stores/AboutOurHoneymoonStore.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import actions from '../actions/AboutOurHoneymoonActions';
 import BaseStore from './BaseStore';
-import { ABOUT_OUR_HONEYMOON as key } from '../constants/keys.constants';
+import { ABOUT_OUR_HONEYMOON as key } from '../constants/KeyConstants';
 
 class AboutOurHoneymoonStore extends BaseStore {
     constructor() {
