+++ public/stores/AboutUsStore.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import actions from '../actions/AboutUsActions';
 import BaseStore from './BaseStore';
-import { ABOUT_US as key } from '../constants/keys.constants';
+import { ABOUT_US as key } from '../constants/KeyConstants';
 
 class AboutUsStore extends BaseStore {
     constructor() {
