+++ public/stores/LocalFlavourStore.js
@@ -1,7 +1,7 @@
 import alt from '../helpers/alt';
 import actions from '../actions/LocalFlavourActions';
 import BaseStore from './BaseStore';
-import { LOCAL_FLAVOUR as key } from '../constants/keys.constants';
+import { LOCAL_FLAVOUR as key } from '../constants/KeyConstants';
 
 class LocalFlavourStore extends BaseStore {
     constructor() {
